refactor: split News into feed terminal + collector config
News promoted to top-level page (pure feed viewer, no tabs).
News Collector added under Data for RSS feed configuration.
