+++ ui/src/App.tsx
@@ -9,12 +9,13 @@ import { SettingsPage } from './pages/SettingsPage'
 import { AIProviderPage } from './pages/AIProviderPage'
 import { MarketDataPage } from './pages/MarketDataPage'
 import { NewsPage } from './pages/NewsPage'
+import { NewsCollectorPage } from './pages/NewsCollectorPage'
 import { TradingPage } from './pages/TradingPage'
 import { ConnectorsPage } from './pages/ConnectorsPage'
 import { DevPage } from './pages/DevPage'
 
 export type Page =
-  | 'chat' | 'portfolio' | 'automation' | 'logs' | 'market-data' | 'news' | 'connectors'
+  | 'chat' | 'portfolio' | 'news' | 'automation' | 'logs' | 'market-data' | 'news-collector' | 'connectors'
   | 'trading'
   | 'ai-provider' | 'settings' | 'dev'
 
@@ -25,6 +26,7 @@ export const ROUTES: Record<Page, string> = {
   'automation': '/automation',
   'logs': '/logs',
   'market-data': '/market-data',
+  'news-collector': '/news-collector',
   'news': '/news',
   'connectors': '/connectors',
   'trading': '/trading',
@@ -66,6 +68,7 @@ export function App() {
             <Route path="/automation" element={<AutomationPage />} />
             <Route path="/logs" element={<LogsPage />} />
             <Route path="/market-data" element={<MarketDataPage />} />
+            <Route path="/news-collector" element={<NewsCollectorPage />} />
             <Route path="/news" element={<NewsPage />} />
             {/* Redirects for old URLs */}
             <Route path="/events" element={<Navigate to="/logs" replace />} />
