+++ ui/src/components/Sidebar.tsx
@@ -46,6 +46,18 @@ const NAV_SECTIONS: NavSection[] = [
           </svg>
         ),
       },
+      {
+        page: 'news',
+        label: 'News',
+        icon: (active) => (
+          <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
+            <path d="M4 22h16a2 2 0 0 0 2-2V4a2 2 0 0 0-2-2H8a2 2 0 0 0-2 2v16a2 2 0 0 1-2 2zm0 0a2 2 0 0 1-2-2v-9h4" />
+            <path d="M10 7h8" />
+            <path d="M10 11h8" />
+            <path d="M10 15h4" />
+          </svg>
+        ),
+      },
     ],
   },
   {
@@ -77,14 +89,13 @@ const NAV_SECTIONS: NavSection[] = [
         ),
       },
       {
-        page: 'news',
-        label: 'News',
+        page: 'news-collector',
+        label: 'News Collector',
         icon: (active) => (
           <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
-            <path d="M4 22h16a2 2 0 0 0 2-2V4a2 2 0 0 0-2-2H8a2 2 0 0 0-2 2v16a2 2 0 0 1-2 2zm0 0a2 2 0 0 1-2-2v-9h4" />
-            <path d="M10 7h8" />
-            <path d="M10 11h8" />
-            <path d="M10 15h4" />
+            <path d="M4 11a9 9 0 0 1 9 9" />
+            <path d="M4 4a16 16 0 0 1 16 16" />
+            <circle cx="5" cy="19" r="1" />
           </svg>
         ),
       },
