+++ ui/src/pages/NewsPage.tsx
@@ -1,9 +1,5 @@
 import { useState, useEffect, useCallback } from 'react'
-import { api, type AppConfig, type NewsCollectorConfig, type NewsCollectorFeed, type NewsArticle } from '../api'
-import { SaveIndicator } from '../components/SaveIndicator'
-import { ConfigSection, Field, inputClass } from '../components/form'
-import { Toggle } from '../components/Toggle'
-import { useConfigPage } from '../hooks/useConfigPage'
+import { api, type NewsArticle } from '../api'
 import { PageHeader } from '../components/PageHeader'
 import { EmptyState } from '../components/StateViews'
 
@@ -24,99 +20,7 @@ const LOOKBACK_OPTIONS = [
   { value: '7d', label: '7 days' },
 ]
 
-// ==================== Feed Section ====================
-
-function NewsFeedSection() {
-  const [articles, setArticles] = useState<NewsArticle[]>([])
-  const [lookback, setLookback] = useState('24h')
-  const [sourceFilter, setSourceFilter] = useState('')
-  const [loading, setLoading] = useState(true)
-  const [sources, setSources] = useState<string[]>([])
-
-  const fetchArticles = useCallback(async (lb: string, src: string) => {
-    try {
-      const res = await api.news.list({
-        lookback: lb,
-        limit: 200,
-        source: src || undefined,
-      })
-      setArticles(res.items)
-      // Collect unique sources for filter dropdown
-      const seen = new Set<string>()
-      for (const item of res.items) {
-        if (item.source) seen.add(item.source)
-      }
-      setSources((prev) => {
-        const merged = new Set([...prev, ...seen])
-        return [...merged].sort()
-      })
-    } catch (err) {
-      console.warn('Failed to load news:', err)
-    } finally {
-      setLoading(false)
-    }
-  }, [])
-
-  // Fetch on mount and when filters change
-  useEffect(() => {
-    setLoading(true)
-    fetchArticles(lookback, sourceFilter)
-  }, [lookback, sourceFilter, fetchArticles])
-
-  // Poll every 60s
-  useEffect(() => {
-    const id = setInterval(() => fetchArticles(lookback, sourceFilter), 60_000)
-    return () => clearInterval(id)
-  }, [lookback, sourceFilter, fetchArticles])
-
-  return (
-    <div className="flex flex-col gap-3 h-full">
-      {/* Controls */}
-      <div className="flex items-center gap-3 shrink-0 flex-wrap">
-        <select
-          value={lookback}
-          onChange={(e) => setLookback(e.target.value)}
-          className="bg-bg-tertiary text-text text-sm rounded-md border border-border px-2 py-1.5 outline-none focus:border-accent"
-        >
-          {LOOKBACK_OPTIONS.map((o) => (
-            <option key={o.value} value={o.value}>{o.label}</option>
-          ))}
-        </select>
-
-        <select
-          value={sourceFilter}
-          onChange={(e) => setSourceFilter(e.target.value)}
-          className="bg-bg-tertiary text-text text-sm rounded-md border border-border px-2 py-1.5 outline-none focus:border-accent"
-        >
-          <option value="">All sources</option>
-          {sources.map((s) => (
-            <option key={s} value={s}>{s}</option>
-          ))}
-        </select>
-
-        <span className="text-xs text-text-muted ml-auto">
-          {articles.length} article{articles.length !== 1 ? 's' : ''}
-        </span>
-      </div>
-
-      {/* Article list */}
-      <div className="flex-1 min-h-0 overflow-y-auto rounded-lg border border-border bg-bg">
-        {loading && articles.length === 0 ? (
-          <div className="px-4 py-8 text-center text-text-muted">Loading...</div>
-        ) : articles.length === 0 ? (
-          <EmptyState title="No articles" description="No news articles found for this time range." />
-        ) : (
-          <div className="divide-y divide-border/50">
-            {/* Reverse to show newest first */}
-            {[...articles].reverse().map((article, i) => (
-              <ArticleRow key={`${article.time}-${i}`} article={article} />
-            ))}
-          </div>
-        )}
-      </div>
-    </div>
-  )
-}
+// ==================== Article Row ====================
 
 function ArticleRow({ article }: { article: NewsArticle }) {
   const [expanded, setExpanded] = useState(false)
@@ -178,198 +82,96 @@ function ArticleRow({ article }: { article: NewsArticle }) {
   )
 }
 
-// ==================== Settings Section ====================
-
-const DEFAULT_NEWS_CONFIG: NewsCollectorConfig = {
-  enabled: true,
-  intervalMinutes: 10,
-  maxInMemory: 2000,
-  retentionDays: 7,
-  feeds: [],
-}
-
-function NewsSettingsSection() {
-  const { config, status, loadError, updateConfig, updateConfigImmediate, retry } = useConfigPage<NewsCollectorConfig>({
-    section: 'news',
-    extract: (full: AppConfig) => (full as Record<string, unknown>).news as NewsCollectorConfig,
-  })
-
-  const cfg = config ?? DEFAULT_NEWS_CONFIG
-  const enabled = cfg.enabled !== false
-
-  return (
-    <div className="flex-1 overflow-y-auto">
-      <div className="flex items-center justify-end gap-3 mb-4">
-        <SaveIndicator status={status} onRetry={retry} />
-        <Toggle size="sm" checked={enabled} onChange={(v) => updateConfigImmediate({ enabled: v })} />
-      </div>
-
-      <div className={`max-w-[880px] mx-auto ${!enabled ? 'opacity-40 pointer-events-none' : ''}`}>
-        {/* Collection Settings */}
-        <ConfigSection
-          title="Collection Settings"
-          description="Control how often articles are fetched and how long they are retained in the archive."
-        >
-          <div className="grid grid-cols-2 gap-4">
-            <Field label="Fetch interval (min)">
-              <input
-                className={inputClass}
-                type="number"
-                min={1}
-                value={cfg.intervalMinutes}
-                onChange={(e) => updateConfig({ intervalMinutes: Number(e.target.value) || 10 })}
-              />
-            </Field>
-            <Field label="Retention (days)">
-              <input
-                className={inputClass}
-                type="number"
-                min={1}
-                value={cfg.retentionDays}
-                onChange={(e) => updateConfig({ retentionDays: Number(e.target.value) || 7 })}
-              />
-            </Field>
-          </div>
-        </ConfigSection>
-
-        {/* RSS Feeds */}
-        <FeedsSection
-          feeds={cfg.feeds}
-          onChange={(feeds) => updateConfigImmediate({ feeds })}
-        />
-      </div>
-      {loadError && <p className="text-[13px] text-red mt-4 max-w-[880px] mx-auto">Failed to load configuration.</p>}
-    </div>
-  )
-}
-
-// ==================== Feeds Section ====================
-
-function FeedsSection({
-  feeds,
-  onChange,
-}: {
-  feeds: NewsCollectorFeed[]
-  onChange: (feeds: NewsCollectorFeed[]) => void
-}) {
-  const [newName, setNewName] = useState('')
-  const [newUrl, setNewUrl] = useState('')
-  const [newSource, setNewSource] = useState('')
-
-  const removeFeed = (index: number) => onChange(feeds.filter((_, i) => i !== index))
-
-  const addFeed = () => {
-    if (!newName.trim() || !newUrl.trim() || !newSource.trim()) return
-    onChange([...feeds, { name: newName.trim(), url: newUrl.trim(), source: newSource.trim() }])
-    setNewName('')
-    setNewUrl('')
-    setNewSource('')
-  }
-
-  return (
-    <ConfigSection
-      title="RSS Feeds"
-      description={
-        feeds.length > 0
-          ? `${feeds.length} feed${feeds.length > 1 ? 's' : ''} configured. Articles are searchable via globNews, grepNews, and readNews tools.`
-          : 'No feeds configured yet. Add feeds to start collecting articles.'
-      }
-    >
-      {/* Existing feeds */}
-      {feeds.length > 0 && (
-        <div className="space-y-2 mb-4">
-          {feeds.map((feed, i) => (
-            <div
-              key={`${feed.source}-${i}`}
-              className="flex items-center gap-3 border border-border/60 rounded-lg px-3 py-2.5"
-            >
-              <div className="flex-1 min-w-0">
-                <p className="text-[13px] font-medium text-text truncate">{feed.name}</p>
-                <p className="text-[12px] text-text-muted truncate">{feed.url}</p>
-                <p className="text-[11px] text-text-muted/50 mt-0.5">source: {feed.source}</p>
-              </div>
-              <button
-                onClick={() => removeFeed(i)}
-                className="shrink-0 text-text-muted hover:text-red transition-colors p-1"
-                title="Remove feed"
-              >
-                <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-                  <line x1="18" y1="6" x2="6" y2="18" />
-                  <line x1="6" y1="6" x2="18" y2="18" />
-                </svg>
-              </button>
-            </div>
-          ))}
-        </div>
-      )}
-
-      {/* Add feed form */}
-      <div className="border border-border/40 rounded-lg p-4 space-y-3">
-        <p className="text-[13px] font-medium text-text-muted">Add Feed</p>
-        <div className="grid grid-cols-2 gap-3">
-          <Field label="Name">
-            <input className={inputClass} value={newName} onChange={(e) => setNewName(e.target.value)} placeholder="e.g. CoinDesk" />
-          </Field>
-          <Field label="Source Tag">
-            <input className={inputClass} value={newSource} onChange={(e) => setNewSource(e.target.value)} placeholder="e.g. coindesk" />
-          </Field>
-        </div>
-        <Field label="Feed URL">
-          <input className={inputClass} value={newUrl} onChange={(e) => setNewUrl(e.target.value)} placeholder="https://example.com/rss.xml" />
-        </Field>
-        <button
-          onClick={addFeed}
-          disabled={!newName.trim() || !newUrl.trim() || !newSource.trim()}
-          className="btn-secondary"
-        >
-          Add Feed
-        </button>
-      </div>
-    </ConfigSection>
-  )
-}
-
 // ==================== Page ====================
 
-type Tab = 'feed' | 'settings'
-
-const TABS: { key: Tab; label: string }[] = [
-  { key: 'feed', label: 'Feed' },
-  { key: 'settings', label: 'Settings' },
-]
-
 export function NewsPage() {
-  const [tab, setTab] = useState<Tab>('feed')
+  const [articles, setArticles] = useState<NewsArticle[]>([])
+  const [lookback, setLookback] = useState('24h')
+  const [sourceFilter, setSourceFilter] = useState('')
+  const [loading, setLoading] = useState(true)
+  const [sources, setSources] = useState<string[]>([])
+
+  const fetchArticles = useCallback(async (lb: string, src: string) => {
+    try {
+      const res = await api.news.list({
+        lookback: lb,
+        limit: 200,
+        source: src || undefined,
+      })
+      setArticles(res.items)
+      const seen = new Set<string>()
+      for (const item of res.items) {
+        if (item.source) seen.add(item.source)
+      }
+      setSources((prev) => {
+        const merged = new Set([...prev, ...seen])
+        return [...merged].sort()
+      })
+    } catch (err) {
+      console.warn('Failed to load news:', err)
+    } finally {
+      setLoading(false)
+    }
+  }, [])
+
+  useEffect(() => {
+    setLoading(true)
+    fetchArticles(lookback, sourceFilter)
+  }, [lookback, sourceFilter, fetchArticles])
+
+  useEffect(() => {
+    const id = setInterval(() => fetchArticles(lookback, sourceFilter), 60_000)
+    return () => clearInterval(id)
+  }, [lookback, sourceFilter, fetchArticles])
 
   return (
     <div className="flex flex-col flex-1 min-h-0">
-      <PageHeader
-        title="News"
-        description="RSS feed collection and article archive."
-      />
-
-      <div className="px-4 md:px-6 border-b border-border/60">
-        <div className="flex gap-1">
-          {TABS.map((t) => (
-            <button
-              key={t.key}
-              onClick={() => setTab(t.key)}
-              className={`px-3 py-2 text-sm font-medium transition-colors relative ${
-                tab === t.key ? 'text-accent' : 'text-text-muted hover:text-text'
-              }`}
-            >
-              {t.label}
-              {tab === t.key && (
-                <div className="absolute bottom-0 left-0 right-0 h-[2px] bg-accent rounded-t" />
-              )}
-            </button>
-          ))}
-        </div>
-      </div>
+      <PageHeader title="News" />
 
       <div className="flex-1 flex flex-col min-h-0 px-4 md:px-6 py-5">
-        <div className="flex-1 min-h-0">
-          {tab === 'feed' ? <NewsFeedSection /> : <NewsSettingsSection />}
+        <div className="flex flex-col gap-3 h-full">
+          {/* Controls */}
+          <div className="flex items-center gap-3 shrink-0 flex-wrap">
+            <select
+              value={lookback}
+              onChange={(e) => setLookback(e.target.value)}
+              className="bg-bg-tertiary text-text text-sm rounded-md border border-border px-2 py-1.5 outline-none focus:border-accent"
+            >
+              {LOOKBACK_OPTIONS.map((o) => (
+                <option key={o.value} value={o.value}>{o.label}</option>
+              ))}
+            </select>
+
+            <select
+              value={sourceFilter}
+              onChange={(e) => setSourceFilter(e.target.value)}
+              className="bg-bg-tertiary text-text text-sm rounded-md border border-border px-2 py-1.5 outline-none focus:border-accent"
+            >
+              <option value="">All sources</option>
+              {sources.map((s) => (
+                <option key={s} value={s}>{s}</option>
+              ))}
+            </select>
+
+            <span className="text-xs text-text-muted ml-auto">
+              {articles.length} article{articles.length !== 1 ? 's' : ''}
+            </span>
+          </div>
+
+          {/* Article list */}
+          <div className="flex-1 min-h-0 overflow-y-auto rounded-lg border border-border bg-bg">
+            {loading && articles.length === 0 ? (
+              <div className="px-4 py-8 text-center text-text-muted">Loading...</div>
+            ) : articles.length === 0 ? (
+              <EmptyState title="No articles" description="No news articles found for this time range." />
+            ) : (
+              <div className="divide-y divide-border/50">
+                {[...articles].reverse().map((article, i) => (
+                  <ArticleRow key={`${article.time}-${i}`} article={article} />
+                ))}
+              </div>
+            )}
+          </div>
         </div>
       </div>
     </div>
